requirements.txt
Flask==3.1.0
python_bcrypt==0.3.2
supabase==2.12.0
